% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelList.R
\name{ModelList}
\alias{ModelList}
\alias{createModelList}
\alias{summary.ModelList}
\alias{predict.ModelList}
\title{ModelList}
\usage{
createModelList(...)

\method{summary}{ModelList}(object, ...)

\method{predict}{ModelList}(object, ...)
}
\arguments{
\item{...}{: fitted models}

\item{object}{: object of class "ModelList"}
}
\value{
an object of class "ModelList"
}
\description{
createModelList() combines fitted models into an object of class "ModelList".

summary.ModelList() is the summary method for class "ModelList".

predict.ModelList() is the predict method for class "ModelList".
}
\examples{
library(dplyr)
dataOld <- gapminder::gapminder \%>\% filter(year <= 2000)
mod1 <- lm(lifeExp ~ year + pop + gdpPercap, dataOld)
mod2 <- rpart::rpart(lifeExp ~ year + pop + gdpPercap, dataOld)
mod3 <- glmnetUtils::cv.glmnet(lifeExp ~ year + pop + gdpPercap, dataOld)
ModList <- createModelList(mod1, mod2, mod3)
summary(ModList)
}
